fix(*): Fix issues with dependency analyzers
Fix issues with css and js dependency analyzers.

Ensure the analyze result object is constructed in the order the
source files are passed to the analyze() method. This has an affect
if array properties are concatenated since each analyzer can finish
at any time, the end result would be non-deterministic.