refactor(*): Clean up changelog and release tools
Refactor changelog tool so that changelogs for other tags can be
generated as well as writing the changelog in Markdown.

Refactor the release tool so that it's ready to create a release
on GitHub or some other release platform.

Install ESLINT and resolve several small issues through out the
tools source code.