feat(*): Refactor recursive analyzers and API
Refactor so that all analyzers are automatically recursive so long
as they return results that adhere to a small API.

For analyzers to be automatically recurively results must either be
a strings that resembles absolute file paths or an objects with the
property 'fileName' that resembles absolute file paths.

NOTE: Absolute file paths typically begin with '/' or 'C:\'
depending on the OS.

BREAKING CHANGE:

Due to analyzers being automatically recursive, all analyzers must
have a 'kind' property specified on the analyze function or be an
object of the shape { kind, analyze }. The kind property is used
to classify analyzers for more efficient recursive analysis.

Before:

Anna
.create()
.use('assets', /.css$/, function (sourceFile, sourceText, anna) {})

After:

Anna.create().use('assets/, /.css$/, {
  kind: 'assets',
  analyze (sourceFile, sourceText, anna) {}
})